PI-2401 Move telemetry out of offence creation transaction
To prevent duplicate entries in the offence report for failures
